convert Disco.Compile to capability style
Also, move some capability-related utilities into Disco.Capability
