+++ disco.cabal
@@ -268,6 +268,7 @@ library
                        Disco.Syntax.Operators
                        Disco.Syntax.Prims
                        Disco.Extensions
+                       Disco.Capability
                        Disco.AST.Core
                        Disco.AST.Desugared
                        Disco.AST.Generic
